Add new files for individual classes in Duke.java
