+++ src/main/java/Duke.java
@@ -8,124 +8,6 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 import java.time.format.DateTimeFormatter;
 
-class User {
-    Scanner sc = new Scanner(System.in);
-    public String command() {
-        if (sc.hasNextLine()) {
-            String str = sc.nextLine();
-            return str;
-        }
-        return "";
-    }
-}
-class Task {
-    protected String description;
-    protected boolean isDone;
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-    public String getStatusIcon() {
-        return (isDone ? "X" : " "); // mark done task with X
-    }
-    public void markAsDone() {
-        this.isDone = true;
-    }
-    public String taskList() {
-        return "|" + " [" + this.getStatusIcon() + "]" + " | " + this.description;
-    }
-    @Override
-    public String toString() {
-        return "[" + this.getStatusIcon() + "]" + " " + this.description;
-    }
-}
-class Deadline extends Task {
-    protected LocalDateTime by;
-    
-    public Deadline(String description, LocalDateTime by) {
-        super(description);
-        this.by = by;
-    }
-    @Override
-    public String taskList() {
-        return "Deadline " + super.taskList() + "(by: " + by + ")";
-    }
-    @Override
-    public String toString() {
-        DateTimeFormatter dtf = DateTimeFormatter.ofPattern("MMM d yyyy, h a");
-        String formattedDtf = this.by.format(dtf);
-        return "[D]" + super.toString() + "(by: " + formattedDtf + ")";
-    }
-}
-class Event extends Task {
-    protected LocalDateTime at;
-    public Event(String description, LocalDateTime at) {
-        super(description);
-        this.at = at;
-    }
-    @Override
-    public String taskList() {
-        return "Event " + super.taskList() + "(at: " + at + ")";
-    }
-    @Override
-    public String toString() {
-        DateTimeFormatter dtf = DateTimeFormatter.ofPattern("MMM d yyyy, h a");
-        String formattedDtf = this.at.format(dtf);
-        return "[E]" + super.toString() + " (at: " + formattedDtf + ")";
-    }
-}
-class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
-    @Override
-    public String taskList() {
-        return "Todo " + super.taskList();
-    }
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
-class DukeException extends Exception {
-    private String message;
-    public DukeException(String message) {
-        this.message = message;
-    }
-    @Override
-    public String toString() {
-        return message;
-    }
-}
-
-class EmptyDescriptionException extends DukeException {
-    public EmptyDescriptionException(String message) {
-        super(message);
-    }
-    @Override
-    public String toString() {
-        return super.toString();
-    }
-}
-
-class InvalidCommandException extends DukeException {
-    public InvalidCommandException(String message) {
-        super(message);
-    }
-    public String toString() {
-        return super.toString();
-    }
-}
-
-class InvalidDateTimeException extends DukeException {
-    public InvalidDateTimeException(String message) {
-        super(message);
-    }
-    public String toString() {
-        return super.toString();
-    }
-}
-
 public class Duke {
     public static void main(String[] args) {
         String logo = " ____        _        \n"
