+++ src/main/java/Task.java
@@ -1,0 +1,27 @@
+public class Task {
+    protected String description;
+    protected boolean isDone;
+    
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+    
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+    
+    public void markAsDone() {
+        this.isDone = true;
+    }
+    
+    public String taskList() {
+        return "|" + " [" + this.getStatusIcon() + "]" + " | " + this.description;
+    }
+    
+    @Override
+    public String toString() {
+        return "[" + this.getStatusIcon() + "]" + " " + this.description;
+    }
+    
+}
