+++ src/main/java/User.java
@@ -1,0 +1,13 @@
+import java.util.Scanner;
+
+public class User {
+    Scanner sc = new Scanner(System.in);
+    
+    public String command() {
+        if (sc.hasNextLine()) {
+            String str = sc.nextLine();
+            return str;
+        }
+        return "";
+    }
+}
